% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.close}
\alias{bgzf.close}
\title{close bgzf file}
\usage{
bgzf.close(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}
}
\value{
0 on success and -1 on error
}
\description{
close bgzf file
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
bgzf.close(fh)
}

