% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.flush}
\alias{bgzf.flush}
\title{flush bgzf file}
\usage{
bgzf.flush(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}
}
\value{
0 on success and -1 on error
}
\description{
flush bgzf file
}

