% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.has.index}
\alias{bgzf.has.index}
\title{tell if index is present}
\usage{
bgzf.has.index(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object}
}
\value{
logical TRUE if index is present
}
\description{
tell if index is present
}
\note{
this function just check for the index presence not its
validity
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
bgzf.has.index(fh)
bgzf.reindex(fh)
bgzf.has.index(fh)
bgzf.close(fh)
}

