% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.index.build.init}
\alias{bgzf.index.build.init}
\title{setup indexing}
\usage{
bgzf.index.build.init(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}
}
\value{
0 on success and -1 on error
}
\description{
this instructs bgzf to build index while compressing or uncompressing file.
}
\note{
this function is mostly used when creating a file for writing. when reading from an existing
file you may consider using \link{bgzf.reindex} instead.
}
\seealso{
\link{bgzf.reindex}, \link{bgzf.index.load}
}

