% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.index.load}
\alias{bgzf.index.load}
\title{load index from file}
\usage{
bgzf.index.load(bgzf, path)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}

\item{path}{path to index file}
}
\value{
0 on success and -1 on error
}
\description{
load index from file
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
bgzf.reindex(fh)
tmp <- tempfile()
bgzf.index.dump(fh, tmp)
bgzf.index.remove(fh)
bgzf.index.load(fh, tmp)
bgzf.useek(fh, 1024)
x <- bgzf.read(fh, 255)
identical(as.integer(x[(1:8)\%\%8==1])-127L, 1:32)
bgzf.close(fh)
unlink(tmp)
}
\seealso{
\link{bgzf.reindex} for a dynamic version. \link{bgzf.index.dump}
}

