% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.index.remove}
\alias{bgzf.index.remove}
\title{remove index}
\usage{
bgzf.index.remove(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}
}
\value{
0 on success and -1 on error
}
\description{
remove index
}

