% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.is.bgzf}
\alias{bgzf.is.bgzf}
\title{check if a file is in the BGZF format}
\usage{
bgzf.is.bgzf(path)
}
\arguments{
\item{path}{file pathname}
}
\value{
1 if file is BGZF; 0 if not or on I/O error
}
\description{
check if a file is in the BGZF format
}
\examples{
path <- file.path(path.package(package="rbgzf"), "samples/test.bin.bgz")
bgzf.is.bgzf(path)
}

