% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.open}
\alias{bgzf.open}
\title{open bgzf file}
\usage{
bgzf.open(path, mode = "r")
}
\arguments{
\item{path}{file pathname}

\item{mode}{open mode, one of "r" or "w"}
}
\value{
BGZFPtr object or NULL on error
}
\description{
open bgzf file
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
bgzf.close(fh)
}

