% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.read}
\alias{bgzf.read}
\title{read raw bytes from bgzf file}
\usage{
bgzf.read(bgzf, nbytes)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}

\item{nbytes}{double number of bytes to read}
}
\value{
raw array of size nbytes (or less if EOF reached)
}
\description{
read raw bytes from bgzf file
}
\note{
to overcome the 32 bits integer limitation, nbytes is interpreted
as a double. this is valid up to 53 bits (i.e 9007199254740992), larger
values will result in incorrect number of bytes read.
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
x <- bgzf.read(fh, 128)
identical(as.integer(x[(1:8)\%\%8==1]), 0:15)
bgzf.close(fh)
}

