% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.reindex}
\alias{bgzf.reindex}
\title{force reindexing whole bgzf file}
\usage{
bgzf.reindex(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}
}
\value{
0 on success and -1 on error
}
\description{
this function (re)index bgzf file by performing the following actions:\cr
\itemize{
 \item remove previous index (if any) and init new index
 \item rewind file to origin
 \item read by blocks (this force indexing while uncompressing)
 \item reposition to initial offset
}
}
\note{
this procedure is quite quick (about 1Gb of uncompressed data per second).
So, for files of moderate size it is therefore not useful to save the
index on disk. For very large files (> 500 Gb) \link{bgzf.index.load} could
be more advantageous.
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
bgzf.reindex(fh)
bgzf.useek(fh, 1024)
x <- bgzf.read(fh, 255)
identical(as.integer(x[(1:8)\%\%8==1])-127L, 1:32)
bgzf.close(fh)
}

