% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.useek}
\alias{bgzf.useek}
\title{position file at uncompressed file offset}
\usage{
bgzf.useek(bgzf, uoffset)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}

\item{uoffset}{offset in uncompressed file}
}
\value{
0 on success and -1 on error
}
\description{
position file at uncompressed file offset
}
\note{
the file should have been indexed (at least up to requested position)

to overcome the 32 bits integer limitation, uoffset is interpreted
as a double. this is valid up to 53 bits (i.e 9007199254740992), larger
values may result in incorrect position.
}
\examples{
fh <- bgzf.open(file.path(path.package(package="rbgzf"), "samples/test.bin.bgz"))
bgzf.reindex(fh)
bgzf.useek(fh, 1024)
x <- bgzf.read(fh, 255)
identical(as.integer(x[(1:8)\%\%8==1])-127L, 1:32)
bgzf.close(fh)
}
\seealso{
\link{bgzf.reindex}, \link{bgzf.index.load}
}

