% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.utell}
\alias{bgzf.utell}
\title{tell byte position in uncompressed file}
\usage{
bgzf.utell(bgzf)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}
}
\value{
the current offset in uncompressed file on success and -1 on error
}
\description{
tell byte position in uncompressed file
}
\note{
the file should have been indexed (at least up to current position)
}
\seealso{
\link{bgzf.reindex}, \link{bgzf.index.load}
}

