% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{bgzf.write}
\alias{bgzf.write}
\title{write raw bytes to bgzf file}
\usage{
bgzf.write(bgzf, data)
}
\arguments{
\item{bgzf}{BGZFPtr object created by \link{bgzf.open}}

\item{data}{array of raw data}
}
\value{
number of bytes written (negative on error)
}
\description{
write raw bytes to bgzf file
}
\examples{
tmp <- tempfile()
fh <- bgzf.open(tmp, "w")
bgzf.write(fh, as.raw((1:1024)\%\%256))
bgzf.flush(fh)
bgzf.close(fh)
fh <- bgzf.open(tmp, "r")
x <- bgzf.read(fh, 255)
identical(as.integer(x), 1:255)
bgzf.close(fh)
unlink(tmp)
}

