% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\name{print.BGZFPtr}
\alias{print.BGZFPtr}
\title{print method for BGZFPtr object}
\usage{
\method{print}{BGZFPtr}(x, ...)
}
\arguments{
\item{x}{a BGZFPtr object}

\item{...}{further arguments passed to or from other methods}
}
\description{
print method for BGZFPtr object
}

