% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzf.r
\docType{package}
\name{rbgzf}
\alias{rbgzf}
\alias{rbgzf-package}
\title{BGZF file format R interface}
\description{
Bgzf (Blocked GNU Zip Format) is a variant of GZip format that allows
direct access into compressed file using offset in the uncompressed file.
The basic idea is to zip file by blocks (of 65K by default) and to maintain
an index between blocks offsets in compressed and uncompressed data.\cr
This R package provides a (thin) interface between R and the bgzf
implementation of the htslib library available at:\cr
\url{https://github.com/samtools/htslib}.
(appropriate LICENSE file is available in the src directory).\cr
The code has been slightly modified to remove dependencies with other
library components (SAM, CRAM, ...).\cr
Note: for WGInR users: WGInR contains an older (but compatible) version of bgzip.
the code for multithreading read operations has unfortunately not (yet) been transfered
from this version (to be done in future).
}
\details{
\tabular{ll}{
Package: \tab rbgzf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-11-02\cr
License: \tab GPL\cr
}
}
\author{
Alain Viari
}

